+++ extras/CreateJobContextForTesting.R
@@ -18,13 +18,45 @@ getSampleCohortDefintionSet <- function() {
     cohortName <- tools::file_path_sans_ext(basename(cohortJsonFileName))
     cohortJson <- readChar(cohortJsonFileName, file.info(cohortJsonFileName)$size)
     sampleCohorts <- rbind(sampleCohorts, data.frame(
-      cohortId = bit64::as.integer64(i),
+      cohortId = as.double(i),
       cohortName = cohortName,
       json = cohortJson,
       sql = "",
       stringsAsFactors = FALSE
     ))
   }
+
+  # Add subsets to the cohort definition set
+  maleOnlySubsetOperators <- list(
+    CohortGenerator::createDemographicSubset(
+      id = 1001,
+      name = "Gender == Male",
+      gender = 8507
+    )
+  )
+  maleOnlySubsetDef <- CohortGenerator::createCohortSubsetDefinition(
+    name = "Males",
+    definitionId = 1,
+    subsetOperators = maleOnlySubsetOperators
+  )
+  # Define a subset for males age 40+
+  maleAgeBoundedSubsetOperators <- list(
+    CohortGenerator::createDemographicSubset(
+      id = 1002,
+      name = "Gender == Male, Age 40+",
+      gender = 8507,
+      ageMin = 40
+    )
+  )
+  maleAgeBoundedSubsetDef <- CohortGenerator::createCohortSubsetDefinition(
+    name = "Male, Age 40+",
+    definitionId = 2,
+    subsetOperators = maleAgeBoundedSubsetOperators
+  )
+
+  sampleCohorts <- sampleCohorts %>%
+    CohortGenerator::addCohortSubsetDefinition(maleOnlySubsetDef) %>%
+    CohortGenerator::addCohortSubsetDefinition(maleAgeBoundedSubsetDef)
   return(sampleCohorts)
 }
 
@@ -33,14 +65,27 @@ createCohortSharedResource <- function(cohortDefinitionSet = getSampleCohortDefi
   return(sharedResource)
 }
 
+createCohortSubsetDefinitionSharedResource <- function(cohortDefinitionSet = getSampleCohortDefintionSet()) {
+  sharedResource <- createCohortSubsetDefinitionSharedResourceSpecifications(cohortDefinitionSet = cohortDefinitionSet)
+  return(sharedResource)
+}
+
+createCohortSubsetSharedResource <- function(cohortDefinitionSet = getSampleCohortDefintionSet()) {
+  sharedResource <- createCohortSubsetSharedResourceSpecifications(cohortDefinitionSet = cohortDefinitionSet)
+  return(sharedResource)
+}
+
 createNegativeControlSharedResource <- function() {
   negativeControlOutcomes <- readCsv(file = system.file("testdata/negativecontrols/negativecontrolOutcomes.csv",
-                                                        package = "CohortGenerator",
-                                                        mustWork = TRUE))
+    package = "CohortGenerator",
+    mustWork = TRUE
+  ))
   negativeControlOutcomes$cohortId <- negativeControlOutcomes$outcomeConceptId
-  createNegativeControlOutcomeCohortSharedResourceSpecifications(negativeControlOutcomeCohortSet = negativeControlOutcomes,
-                                                                 occurrenceType = "all",
-                                                                 detectOnDescendants = FALSE)
+  createNegativeControlOutcomeCohortSharedResourceSpecifications(
+    negativeControlOutcomeCohortSet = negativeControlOutcomes,
+    occurrenceType = "all",
+    detectOnDescendants = FALSE
+  )
 }
 
 # Create CohortGeneratorModule settings ---------------------------------------
@@ -52,11 +97,13 @@ cohortGeneratorModuleSpecifications <- createCohortGeneratorModuleSpecifications
 # Module Settings Spec ----------------------------
 analysisSpecifications <- createEmptyAnalysisSpecificiations() %>%
   addSharedResources(createCohortSharedResource()) %>%
+  addSharedResources(createCohortSubsetDefinitionSharedResource()) %>%
+  addSharedResources(createCohortSubsetSharedResource()) %>%
   addSharedResources(createNegativeControlSharedResource()) %>%
   addModuleSpecifications(cohortGeneratorModuleSpecifications)
 
-#executionSettings <- Strategus::createExecutionSettings(
-executionSettings <-   Strategus::createCdmExecutionSettings(
+# executionSettings <- Strategus::createExecutionSettings(
+executionSettings <- Strategus::createCdmExecutionSettings(
   connectionDetailsReference = "dummy",
   workDatabaseSchema = "main",
   cdmDatabaseSchema = "main",
@@ -79,3 +126,4 @@ jobContext <- list(
   moduleExecutionSettings = moduleExecutionSettings
 )
 saveRDS(jobContext, "tests/testJobContext.rds")
+# ParallelLogger::saveSettingsToJson(analysisSpecifications, fileName = "extras/analysisSettings.json")