+++ extras/SettingsFunctionsStub.R
@@ -1,4 +1,4 @@
-# Copyright 2022 Observational Health Data Sciences and Informatics
+# Copyright 2023 Observational Health Data Sciences and Informatics
 #
 # This file is part of CohortGeneratorModule
 #
@@ -17,9 +17,9 @@
 #' Create specifications for the CohortGeneratorModule
 #'
 #' @param incremental Should the CohortGenerator module run in incremental mode?
-#'                    
+#'
 #' @param generateStats Should the CohortGenerator module generate cohort statistics?
-#'                    
+#'
 #'
 #' @return
 #' An object of type `CohortGeneratorModuleSpecifications`.
@@ -31,12 +31,14 @@ createCohortGeneratorModuleSpecifications <- function(incremental = TRUE,
   for (name in names(formals(createCohortGeneratorModuleSpecifications))) {
     analysis[[name]] <- get(name)
   }
-  
-  specifications <- list(module = "%module%",
-                         version = "%version%",
-                         remoteRepo = "github.com",
-                         remoteUsername = "ohdsi",
-                         settings = analysis)
+
+  specifications <- list(
+    module = "%module%",
+    version = "%version%",
+    remoteRepo = "github.com",
+    remoteUsername = "ohdsi",
+    settings = analysis
+  )
   class(specifications) <- c("CohortGeneratorModuleSpecifications", "ModuleSpecifications")
   return(specifications)
 }
@@ -54,28 +56,114 @@ createCohortSharedResourceSpecifications <- function(cohortDefinitionSet) {
   if (!CohortGenerator::isCohortDefinitionSet(cohortDefinitionSet)) {
     stop("cohortDefinitionSet is not properly defined")
   }
-  cohortDefinitionSet <- cohortDefinitionSet[,c("cohortId", "cohortName", "json")]
-  names(cohortDefinitionSet) <- c("cohortId", "cohortName", "cohortDefinition")
-  print(cohortDefinitionSet[,c("cohortId", "cohortName")])
-  cohortDefinitionSet <- apply(cohortDefinitionSet, 1, as.list)
-  sharedResource <- list(cohortDefinitions = cohortDefinitionSet)
+
+  subsetDefinitions <- CohortGenerator::getSubsetDefinitions(cohortDefinitionSet)
+  if (length(subsetDefinitions) > 0) {
+    # Filter the cohort definition set to the "parent" cohorts.
+    cohortDefinitionSet <- cohortDefinitionSet[!cohortDefinitionSet$isSubset, ]
+  }
+
+  listafy <- function(df) {
+    mylist <- list()
+    for (i in 1:nrow(df)) {
+      cohortData <- list(
+        cohortId = df$cohortId[i],
+        cohortName = df$cohortName[i],
+        cohortDefinition = df$json[i]
+      )
+      if (i == 1) {
+        mylist <- list(cohortData)
+      } else {
+        mylist <- append(mylist, list(cohortData))
+      }
+    }
+    return(mylist)
+  }
+
+  cohortDefinitionSetFiltered <- listafy(cohortDefinitionSet)
+  sharedResource <- list(cohortDefinitions = cohortDefinitionSetFiltered)
+
   class(sharedResource) <- c("CohortDefinitionSharedResources", "SharedResources")
+  return(sharedResource)
+}
+
+#' Create shared specifications for the cohort subset definitions
+#'
+#' @param cohortDefinitionSet The cohortDefintionSet information about subsets if specified.
+#'
+#' @return
+#' An object of type `CohortSubsetDefinitionSharedResources`.
+#'
+#' @export
+createCohortSubsetDefinitionSharedResourceSpecifications <- function(cohortDefinitionSet) {
+  if (!CohortGenerator::isCohortDefinitionSet(cohortDefinitionSet)) {
+    stop("cohortDefinitionSet is not properly defined")
+  }
+
+  sharedResource <- list()
+  subsetDefinitions <- CohortGenerator::getSubsetDefinitions(cohortDefinitionSet)
+  if (length(subsetDefinitions)) {
+    subsetDefinitionsJson <- lapply(subsetDefinitions, function(x) {
+      x$toJSON()
+    })
+    sharedResource <- list(subsetDefs = subsetDefinitionsJson)
+  }
+
+  class(sharedResource) <- c("CohortSubsetDefinitionSharedResources", "SharedResources")
+  return(sharedResource)
+}
+
+#' Create shared specifications for the cohort subsets
+#'
+#' @param cohortDefinitionSet The cohortDefintionSet information about subsets if specified.
+#'
+#' @return
+#' An object of type `CohortSubsetSharedResources`.
+#'
+#' @export
+createCohortSubsetSharedResourceSpecifications <- function(cohortDefinitionSet) {
+  if (!CohortGenerator::isCohortDefinitionSet(cohortDefinitionSet)) {
+    stop("cohortDefinitionSet is not properly defined")
+  }
+
+  sharedResource <- list()
+  subsetDefinitions <- CohortGenerator::getSubsetDefinitions(cohortDefinitionSet)
+  if (length(subsetDefinitions)) {
+    cohortDefinitionSet <- cohortDefinitionSet[cohortDefinitionSet$isSubset, ]
+
+    subsetIdMapping <- list()
+    for (i in 1:nrow(cohortDefinitionSet)) {
+      idMapping <- list(
+        cohortId = cohortDefinitionSet$cohortId[i],
+        subsetId = cohortDefinitionSet$subsetDefinitionId[i],
+        targetCohortId = cohortDefinitionSet$subsetParent[i]
+      )
+      if (i == 1) {
+        subsetIdMapping <- list(idMapping)
+      } else {
+        subsetIdMapping <- append(subsetIdMapping, list(idMapping))
+      }
+    }
+    sharedResource <- list(cohortSubsets = subsetIdMapping)
+  }
+
+  class(sharedResource) <- c("CohortSubsetSharedResources", "SharedResources")
   return(sharedResource)
 }
 
 #' Create shared specifications for the negative control outcome
 #' cohort set
 #'
-#' @param negativeControlOutcomeCohortSet	The negativeControlOutcomeCohortSet argument 
+#' @param negativeControlOutcomeCohortSet	The negativeControlOutcomeCohortSet argument
 #' must be a data frame with the following columns: cohortId, cohortName, outcomeConceptId
-#' 
-#' @param occurrenceType The occurrenceType will detect either: the first time an 
-#'                       outcomeConceptId occurs or all times the outcomeConceptId 
+#'
+#' @param occurrenceType The occurrenceType will detect either: the first time an
+#'                       outcomeConceptId occurs or all times the outcomeConceptId
 #'                       occurs for a person. Values accepted: 'all' or 'first'.
-#' 
-#' @param detectOnDescendants When set to TRUE, detectOnDescendants will use the vocabulary 
-#'                            to find negative control outcomes using the outcomeConceptId and all 
-#'                            descendants via the concept_ancestor table. When FALSE, only the exact 
+#'
+#' @param detectOnDescendants When set to TRUE, detectOnDescendants will use the vocabulary
+#'                            to find negative control outcomes using the outcomeConceptId and all
+#'                            descendants via the concept_ancestor table. When FALSE, only the exact
 #'                            outcomeConceptId will be used to detect the outcome.
 #'
 #' @return
@@ -90,7 +178,8 @@ createNegativeControlOutcomeCohortSharedResourceSpecifications <- function(negat
     negativeControlOutcomes = list(
       negativeControlOutcomeCohortSet = negativeControlOutcomeCohortSet,
       occurrenceType = occurrenceType,
-      detectOnDescendants = detectOnDescendants)
+      detectOnDescendants = detectOnDescendants
+    )
   )
   class(sharedResource) <- c("NegativeControlOutcomeSharedResources", "SharedResources")
   return(sharedResource)